fastapi/requirements.txt
# FastAPI backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
duckdb==0.10.2
pandas>=2.0.0
pydantic>=2.0.0
boto3>=1.28.0
python-multipart>=0.0.6
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.11.12
cffi==2.0.0
minio==7.2.18
pycparser==2.23
pycryptodome==3.23.0
typing_extensions==4.15.0
urllib3==2.5.0
python-jose
bcrypt
pysqlite3
